refactor(test): the two source scans share one test-module splitter
`non_test_prefix` was copied byte for byte into both scan binaries and had
already drifted once; it now lives in tests/common.
