+++ rust/tests/common/mod.rs
@@ -267,3 +267,25 @@ fn cache_base() -> PathBuf {
     let home = std::env::var("HOME").unwrap_or_else(|_| "/tmp".into());
     PathBuf::from(home).join(".cache/kesha")
 }
+
+/// Source text before the first `#[cfg(test)] mod` or `#[cfg(all(test, …))] mod`; an attribute on a single item would otherwise un-scan everything after it.
+pub fn non_test_prefix(text: &str) -> &str {
+    const MARKERS: [&str; 2] = ["#[cfg(test)]", "#[cfg(all(test,"];
+    let mut from = 0;
+    loop {
+        let Some((at, marker)) = MARKERS
+            .iter()
+            .filter_map(|m| text[from..].find(m).map(|i| (from + i, *m)))
+            .min_by_key(|(i, _)| *i)
+        else {
+            return text;
+        };
+        let Some(close) = text[at..].find(']') else {
+            return text;
+        };
+        if text[at + close + 1..].trim_start().starts_with("mod ") {
+            return &text[..at];
+        }
+        from = at + marker.len();
+    }
+}
